+++ packages/components/chat/chat-action-props.ts
@@ -22,6 +22,7 @@ export default {
   /** 操作按钮配置项，可配置操作按钮选项和顺序 */
   operationBtn: {
     type: Array as PropType<TdChatActionProps['operationBtn']>,
+    default: (): TdChatActionProps['operationBtn'] => ['replay', 'copy', 'good', 'bad'],
   },
   /** 点击点赞，点踩，复制，重新生成按钮时触发 */
   onOperation: Function as PropType<TdChatActionProps['onOperation']>,