+++ packages/components/chat/chat-input-props.ts
@@ -24,6 +24,23 @@ export default {
   },
   /** 中止按钮是否可点击。等流式数据全部返回结束置为false，注意跟textLoading的控制时机不是同一个 */
   stopDisabled: Boolean,
+  /** 发送按钮的自定义扩展 */
+  suffixIcon: {
+    type: Function as PropType<TdChatInputProps['suffixIcon']>,
+  },
+  /** 输入框的值 */
+  value: {
+    type: [String, Number] as PropType<TdChatInputProps['value']>,
+    default: undefined as TdChatInputProps['value'],
+  },
+  modelValue: {
+    type: [String, Number] as PropType<TdChatInputProps['value']>,
+    default: undefined as TdChatInputProps['value'],
+  },
+  /** 输入框的值，非受控属性 */
+  defaultValue: {
+    type: [String, Number] as PropType<TdChatInputProps['defaultValue']>,
+  },
   /** 输入框聚焦时触发 */
   onBlur: Function as PropType<TdChatInputProps['onBlur']>,
   /** 输入框值发生变化时触发 */